+++ modules/runtime-fields-common/build.gradle
@@ -1,0 +1,22 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0 and the Server Side Public License, v 1; you may not use this file except
+ * in compliance with, at your election, the Elastic License 2.0 or the Server
+ * Side Public License, v 1.
+ */
+
+esplugin {
+  description 'Module for runtime fields features and extensions that have large dependencies'
+  classname 'org.elasticsearch.runtimefields.RuntimeFieldsCommonPlugin'
+  extendedPlugins = ['lang-painless']
+}
+
+dependencies {
+  compileOnly project(':modules:lang-painless:spi')
+  api project(':libs:elasticsearch-grok')
+  api project(':libs:elasticsearch-dissect')
+}
+
+//this plugin is here only for the painless extension, there are no unit tests
+tasks.named("testingConventions").configure { enabled = false }
