+++ modules/runtime-fields-common/src/main/resources/org/elasticsearch/xpack/runtimefields/mapper/common_whitelist.txt
@@ -1,8 +1,9 @@
 #
 # Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
 # or more contributor license agreements. Licensed under the Elastic License
-# 2.0; you may not use this file except in compliance with the Elastic License
-# 2.0.
+# 2.0 and the Server Side Public License, v 1; you may not use this file except
+# in compliance with, at your election, the Elastic License 2.0 or the Server
+# Side Public License, v 1.
 #
 
 class org.elasticsearch.xpack.runtimefields.mapper.NamedGroupExtractor @no_import {
