+++ x-pack/plugin/runtime-fields/build.gradle
@@ -4,15 +4,12 @@ esplugin {
   name 'x-pack-runtime-fields'
   description 'A module which adds support for runtime fields'
   classname 'org.elasticsearch.xpack.runtimefields.RuntimeFields'
-  extendedPlugins = ['x-pack-core', 'lang-painless']
+  extendedPlugins = ['x-pack-core']
 }
 archivesBaseName = 'x-pack-runtime-fields'
 
 dependencies {
   compileOnly project(":server")
-  compileOnly project(':modules:lang-painless:spi')
-  api project(':libs:elasticsearch-grok')
-  api project(':libs:elasticsearch-dissect')
   compileOnly project(path: xpackModule('core'))
 }
