--- x-pack/plugin/runtime-fields/src/main/resources/META-INF/services/org.elasticsearch.painless.spi.PainlessExtension
@@ -1,1 +1,0 @@
-org.elasticsearch.xpack.runtimefields.mapper.RuntimeFieldsPainlessExtension
